# docs/schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClinicalTrials.gov Query-to-Visualization Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/query": {
      "post": {
        "summary": "Answer a question with a visualization specification",
        "description": "Plan the question, run it against ClinicalTrials.gov, answer with a chart, a question or a message.",
        "operationId": "query_v1_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/analyses": {
      "post": {
        "summary": "Run a typed plan, with no model call",
        "description": "The same response as `/v1/query` for a plan that is already written: the replay path.",
        "operationId": "analyses_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/schema/{name}": {
      "get": {
        "summary": "A JSON Schema of the contract",
        "description": "`contract`, `query-request`, `query-plan`, `query-response` or `error-response`.",
        "operationId": "get_schema_v1_schema__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/examples/{slug}": {
      "get": {
        "summary": "One recorded run: request, plan and response",
        "operationId": "get_example_v1_examples__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnalysisRequest": {
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/QueryPlan"
          },
          "options": {
            "$ref": "#/components/schemas/RequestOptions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan"
        ],
        "title": "AnalysisRequest",
        "description": "A typed plan to answer without a model: the body of `POST /v1/analyses`.\n\nNothing is checked against a question; the plan is validated for shape and limits only. Posting\n`{\"plan\": meta.plan, \"options\": meta.options}` from an earlier response reproduces its visualization\nfor the same data timestamp."
      },
...
      "ClarificationResponse": {
        "properties": {
          "spec_version": {
            "type": "string",
            "const": "1.0",
            "title": "Spec Version",
            "default": "1.0"
          },
          "kind": {
            "type": "string",
            "const": "clarification",
            "title": "Kind",
            "default": "clarification"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "visualization": {
            "type": "null",
            "title": "Visualization"
          },
          "clarification": {
            "$ref": "#/components/schemas/Clarification"
          },
          "references": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TrialReference"
            },
            "type": "object",
            "title": "References"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "spec_version",
          "kind",
          "message",
          "visualization",
          "clarification",
          "references",
          "meta"
        ],
        "title": "ClarificationResponse",
        "description": "The service needs a name or a choice before it can answer."
      },
...
      "CompareSpec": {
        "properties": {
          "field": {
            "type": "string",
            "enum": [
              "drug_name",
              "condition",
              "sponsor",
              "country"
            ],
            "title": "Field"
          },
          "values": {
            "items": {
              "type": "string",
              "maxLength": 200,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 2,
            "title": "Values"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "values"
        ],
        "title": "CompareSpec",
        "description": "Values of one entity kind to compare side by side, one series each."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "spec_version": {
            "type": "string",
            "const": "1.0",
            "title": "Spec Version",
            "default": "1.0"
          },
          "kind": {
            "type": "string",
            "enum": [
              "no_data",
              "unsupported"
            ],
            "title": "Kind"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "visualization": {
            "type": "null",
            "title": "Visualization"
          },
          "clarification": {
            "type": "null",
            "title": "Clarification"
          },
          "references": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TrialReference"
            },
            "type": "object",
            "title": "References"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "spec_version",
          "kind",
          "message",
          "visualization",
          "clarification",
          "references",
          "meta"
        ],
        "title": "MessageResponse",
        "description": "Nothing to draw: `no_data` when the plan ran and matched nothing, `unsupported` when it cannot run."
      },
...
      "QueryPlan": {
        "properties": {
          "interpretation": {
            "type": "string",
            "title": "Interpretation",
            "description": "One sentence restating what will be counted and how it is grouped. No figures other than those in the question."
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/Entity"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Entities",
            "description": "Named drugs, conditions, sponsors, countries or other terms."
          },
          "filters": {
            "$ref": "#/components/schemas/PlanFilters"
          },
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Aggregate"
              },
              {
                "$ref": "#/components/schemas/Total"
              },
              {
                "$ref": "#/components/schemas/Relate"
              },
              {
                "$ref": "#/components/schemas/Network"
              },
              {
                "$ref": "#/components/schemas/TrialList"
              },
              {
                "$ref": "#/components/schemas/Clarify"
              },
              {
                "$ref": "#/components/schemas/Unsupported"
              }
            ],
            "title": "Analysis"
          },
          "chart_preference": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "bar_chart",
                  "time_series",
                  "histogram",
                  "scatter_plot",
                  "network_graph",
                  "table",
                  "metric"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Chart Preference",
            "description": "Only when the user names a chart form; otherwise null."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "interpretation",
          "entities",
          "filters",
          "analysis",
          "chart_preference"
        ],
        "title": "QueryPlan"
      },
      "QueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Query",
            "description": "The natural-language question. Tabs and line breaks become spaces and runs of whitespace collapse before validation, because a pasted question can carry line breaks. Then 3 to 1,000 characters, at least one letter and no other control characters."
          },
          "drug_name": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 200,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 5
              },
              {
                "type": "null"
              }
            ],
            "title": "Drug Name",
            "description": "Drugs the trials must test, up to 5, each 1 to 200 characters; a single string is accepted. A trial must match all of them. Searched with the registry's intervention search, so brand and code names resolve (Keytruda and MK-3475 return the same trials as pembrolizumab)."
          },
          "condition": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 200,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 5
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition",
            "description": "Conditions the trials must study; same rules as `drug_name`, searched with the registry's condition search."
          },
          "sponsor": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 200,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 5
              },
              {
                "type": "null"
              }
            ],
            "title": "Sponsor",
            "description": "Lead-sponsor names; same rules as `drug_name`, searched in the lead sponsor name."
          },
          "country": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 200,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 5
              },
              {
                "type": "null"
              }
            ],
            "title": "Country",
            "description": "Countries; same rules as `drug_name`. Each must resolve to one of the registry's country names (common aliases and ISO alpha-2 or alpha-3 codes are accepted), otherwise the request is refused with close matches. A trial must list a site in every named country."
          },
          "term": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 200,
                  "minLength": 1
                },
                "type": "array",
                "maxItems": 5
              },
              {
                "type": "null"
              }
            ],
            "title": "Term",
            "description": "Other search words, matched anywhere in the record; same rules as `drug_name`."
          },
          "trial_phase": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "EARLY_PHASE1",
                    "PHASE1",
                    "PHASE2",
                    "PHASE3",
                    "PHASE4",
                    "NA"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Phase",
            "description": "Phases, any of. Lenient spellings are normalised before validation: 'Phase 2', 'phase2', '2' and 'II' become PHASE2, and 'Phase 2/3' becomes both PHASE2 and PHASE3; a single value is accepted. 'Lists this phase' semantics, as on the registry's website: PHASE2 also matches Phase 1/Phase 2 and Phase 2/Phase 3 trials."
          },
          "status": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "NOT_YET_RECRUITING",
                    "RECRUITING",
                    "ENROLLING_BY_INVITATION",
                    "ACTIVE_NOT_RECRUITING",
                    "COMPLETED",
                    "SUSPENDED",
                    "TERMINATED",
                    "WITHDRAWN",
                    "AVAILABLE",
                    "NO_LONGER_AVAILABLE",
                    "TEMPORARILY_NOT_AVAILABLE",
                    "APPROVED_FOR_MARKETING",
                    "WITHHELD",
                    "UNKNOWN"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "Study-level overall statuses, any of."
          },
          "study_type": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "INTERVENTIONAL",
                    "OBSERVATIONAL",
                    "EXPANDED_ACCESS"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Study Type",
            "description": "Study types, any of."
          },
          "sponsor_class": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "NIH",
                    "FED",
                    "OTHER_GOV",
                    "INDIV",
                    "INDUSTRY",
                    "NETWORK",
                    "AMBIG",
                    "OTHER",
                    "UNKNOWN"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sponsor Class",
            "description": "Classes of the lead sponsor, any of."
          },
          "intervention_type": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "BEHAVIORAL",
                    "BIOLOGICAL",
                    "COMBINATION_PRODUCT",
                    "DEVICE",
                    "DIAGNOSTIC_TEST",
                    "DIETARY_SUPPLEMENT",
                    "DRUG",
                    "GENETIC",
                    "PROCEDURE",
                    "RADIATION",
                    "OTHER"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intervention Type",
            "description": "Trials with at least one intervention of these types, any of."
          },
          "start_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Year",
            "description": "First year, inclusive, of the date named by `date_field`. Not after `end_year`."
          },
          "end_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2100.0,
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "End Year",
            "description": "Last year, inclusive, of the date named by `date_field`. Not before `start_year`."
          },
          "date_field": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "start_date",
                  "primary_completion_date",
                  "completion_date",
                  "first_posted_date"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Field",
            "description": "Which date the year bounds apply to. Null means the date on the time axis when the chart has one, otherwise the study start date."
          },
          "compare": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompareSpec"
              },
              {
                "type": "null"
              }
            ],
            "description": "Compare these as series. Replaces any comparison found in the question."
          },
          "group_by": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "phase",
                    "overall_status",
                    "study_type",
                    "sponsor_class",
                    "intervention_type",
                    "sex",
                    "age_group",
                    "allocation",
                    "masking",
                    "primary_purpose",
                    "has_results",
                    "country",
                    "sponsor",
                    "drug",
                    "condition",
                    "start_date",
                    "primary_completion_date",
                    "completion_date",
                    "first_posted_date",
                    "enrollment"
                  ]
                },
                "type": "array",
                "maxItems": 2,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Group By",
            "description": "Analysis hint: the first key is the axis, the second the series. The second key must be one of the closed dimensions (phase, overall_status, study_type, sponsor_class, intervention_type, sex, age_group, allocation, masking, primary_purpose, has_results). Overrides what the planner chose. Required when `options.planner` is 'structured'."
          },
          "time_unit": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "year",
                  "quarter",
                  "month"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Unit",
            "description": "Analysis hint: bucket width for a date dimension."
          },
          "top_n": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 50.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top N",
            "description": "How many categories or rows to keep. Default 15, or 10 for the rows of a trial list. Network sizes are fixed in v1."
          },
          "chart_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "bar_chart",
                  "time_series",
                  "histogram",
                  "scatter_plot",
                  "network_graph",
                  "table",
                  "metric"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Chart Type",
            "description": "A preference. Honoured when valid for the data shape, otherwise ignored with a warning."
          },
          "options": {
            "$ref": "#/components/schemas/RequestOptions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryRequest",
        "description": "A question about clinical trials, with optional structured fields.\n\nStrings are trimmed and unknown keys are rejected. Rules across fields:\n\n1. Structured filter fields are ANDed with what the question says. Where they conflict, the field\n   wins: it replaces the planner's value for the same entity kind or filter family, and the\n   override is recorded in `meta.interpretation.adjustments`.\n2. A structured entity field whose value is one of the compared values does not become an extra\n   filter; the comparison is kept.\n3. Entity arrays mean \"all of\"; enum arrays mean \"any of\". A comparison is always explicit:\n   `compare` in the request or an \"A vs B\" in the question.\n4. `group_by`, `time_unit`, `top_n` and `chart_type` replace the planner's choices.\n5. Every structured field is shown to the model, so \"this drug\" in the question resolves to\n   `drug_name`.\n6. `meta.filters` in every response repeats the filter fields in canonical form (arrays for list\n   fields, every key present), so it can be sent back as request fields. For that reason every list\n   field accepts an empty array as \"no filter\", the same as null.",
        "examples": [
          {
            "drug_name": "Pembrolizumab",
            "query": "How has the number of trials for this drug changed over time?"
          }
        ]
      },
      "QueryResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/VisualizationResponse"
          },
          {
            "$ref": "#/components/schemas/ClarificationResponse"
          },
          {
            "$ref": "#/components/schemas/MessageResponse"
          }
        ],
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "clarification": "#/components/schemas/ClarificationResponse",
            "no_data": "#/components/schemas/MessageResponse",
            "unsupported": "#/components/schemas/MessageResponse",
            "visualization": "#/components/schemas/VisualizationResponse"
          }
        }
      },
...
      "RequestOptions": {
        "properties": {
          "planner": {
            "type": "string",
            "enum": [
              "llm",
              "structured"
            ],
            "title": "Planner",
            "description": "'llm': the model writes the plan (503 `planner_unavailable` when no model is configured). 'structured': never call a model; the plan is built from the request fields and `group_by` is required. The question text is then not interpreted, and the response says so. Ignored by `POST /v1/analyses`.",
            "default": "llm"
          },
          "citations_per_datum": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Citations Per Datum",
            "description": "Trials cited for each datum; 0 disables citations.",
            "default": 5
          },
          "drug_match": {
            "type": "string",
            "enum": [
              "broad",
              "name_only"
            ],
            "title": "Drug Match",
            "description": "'broad': the registry's intervention search (names, other names, titles, descriptions, synonyms). 'name_only': intervention names and their synonyms only. The two give different counts, so the definition used is always stated in the response.",
            "default": "broad"
          },
          "include_trace": {
            "type": "boolean",
            "title": "Include Trace",
            "description": "Include the step list in `meta.debug.trace`.",
            "default": true
          },
          "use_cache": {
            "type": "boolean",
            "title": "Use Cache",
            "description": "false bypasses the plan cache and the response cache for this request; the registry-call cache stays.",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RequestOptions",
        "description": "Behaviour switches of one request; `meta.options` echoes the effective values."
      },
...
      "Total": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "total",
            "title": "Kind"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "Total"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Visualization": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/BarChart"
          },
          {
            "$ref": "#/components/schemas/TimeSeries"
          },
          {
            "$ref": "#/components/schemas/Histogram"
          },
          {
            "$ref": "#/components/schemas/ScatterPlot"
          },
          {
            "$ref": "#/components/schemas/NetworkGraph"
          },
          {
            "$ref": "#/components/schemas/Table"
          },
          {
            "$ref": "#/components/schemas/Metric"
          }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "bar_chart": "#/components/schemas/BarChart",
            "histogram": "#/components/schemas/Histogram",
            "metric": "#/components/schemas/Metric",
            "network_graph": "#/components/schemas/NetworkGraph",
            "scatter_plot": "#/components/schemas/ScatterPlot",
            "table": "#/components/schemas/Table",
            "time_series": "#/components/schemas/TimeSeries"
          }
        }
      },
      "VisualizationResponse": {
        "properties": {
          "spec_version": {
            "type": "string",
            "const": "1.0",
            "title": "Spec Version",
            "default": "1.0"
          },
          "kind": {
            "type": "string",
            "const": "visualization",
            "title": "Kind",
            "default": "visualization"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "A one-sentence headline computed from the data."
          },
          "visualization": {
            "$ref": "#/components/schemas/Visualization"
          },
          "clarification": {
            "type": "null",
            "title": "Clarification"
          },
          "references": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TrialReference"
            },
            "type": "object",
            "title": "References",
            "description": "Exactly the trials cited in the data, keyed by NCT ID."
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "spec_version",
          "kind",
          "message",
          "visualization",
          "clarification",
          "references",
          "meta"
        ],
        "title": "VisualizationResponse",
        "description": "An answer drawn as a chart, a number or a table."
      }
    }
  }
}